#87 refactor: 메시지만 변경되는 토스트 handler 훅 추가
